pyproject.toml
[project]
name = "pyxelart"
version = "0.1.0"
description = "A suite of tools for creating retro-style images, videos and sprites"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=11.2.1",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "tqdm>=4.67.1",
    "rembg>=2.0.65",
    "imageio>=2.37.0",
    "scipy>=1.15.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]

